# docs/swagger.yaml
swagger: '2.0'
info:
  title: Somosport Core API
  description: Basic description of services provided by Somosport Core. Online editor & validator available at  http://editor.swagger.io/
  version: "1.0"
# the domain of the service
# host: ss-core-dev.herokuapp.com
#host: localhost:3000
# array of all schemes that your API supports
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /api/v1.0
produces:
  - application/json

##################################################
#
#  _____                 _
# /  ___|               (_)
# \ `--.  ___ _ ____   ___  ___ ___  ___
#  `--. \/ _ \ '__\ \ / / |/ __/ _ \/ __|
# /\__/ /  __/ |   \ V /| | (_|  __/\__ \
# \____/ \___|_|    \_/ |_|\___\___||___/
#
#
###################################################
paths:
...
  /category/{category_id}/match:
    get:
      summary: Returns the matches associated to to this category
      description: Returns the matches associated to to this category
      tags:
        - Category
        - Match
      responses:
        200:
          description: Something
      parameters:
        - name: category_id
          in: path
          description: Category ID
          required: true
          type: integer
          format: integer
    post:
      summary: Generates the matches for the given category id
      description: Generates the matches for the given category id
      tags:
        - Category
        - Match
      responses:
        200:
          description: Something
      parameters:
        - name: category_id
          in: path
          description: Category ID
          required: true
          type: integer
          format: integer
...
  /category/{category_id}/teams/{team_id}/matches:
    get:
      summary: Returns the matches of a given team, under a category
      tags:
        - Category
        - Team
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
          format: integer
        - name: team_id
          in: path
          required: true
          type: integer
          format: integer
      responses:
        200:
          description: List of matches given a category and a team
...
  /phase/{phase_id}/match:
    get:
      summary: Given a Phase ID, returns the groups & matches associated to that phase
      description: Given a Phase ID, returns the groups & matches associated to that phase
      tags:
        - Phase
        - Match
      responses:
        200:
          description: List of groups & matches for the given phase
      parameters:
        - name: phase_id
          in: path
          description: Phase ID
          required: true
          type: integer
          format: integer
...
  /group/{group_id}/match:
    get:
      description: Get matches of this group
      tags:
        - Group
        - Match
      produces:
      - application/json
      parameters:
        - name: group_id
          in: path
          description: Group ID
          required: true
          type: integer
          format: integer
      responses:
        200:
          description: list of matches currently on this group
      security: []
  /match:
    get:
      summary: Gets all matches
      description: Gets all matches
      tags:
        - Match
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: list of all matches
      security: []
    post:
      summary: Creates a match
      description: Creates a match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/CreateMatchRequest'
      responses:
        200:
          description: match creation OK
      security: []
  /match/{match_id}:
    get:
      summary: Gets one match
      description: Gets one match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: list of all matches
      security: []
    put:
      summary: Updates a match
      description: Updates a match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/CreateMatchRequest'
      responses:
        200:
          description: match creation OK
      security: []
  /match/{match_id}/feed:
    get:
      summary: Gets the feed of one match
      description: Gets the feed of one match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: list of all feed items for this match
      security: []
  /match/{match_id}/player:
    get:
      summary: Returns the player list for this match, grouped by home_team & visitor_team
      description: Returns the player list for this match, grouped by home_team & visitor_team
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: list of players available for this match
  /match/{match_id}/team/{team_id}/player:
    get:
      summary: Gets the player list for this team & match
      description: Gets the player list for this team & match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: List of players assigned to this match & team
    put:
      summary: Toggles the participation of the given player in this match
      description: Toggles the participation of the given player in this match; necessary for updating the summoned player list for each match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: player_data
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/MatchPlayerTeamRequest'
      responses:
        200:
          description: Correct update of match player list
  /match/{match_id}/team:
    get:
      summary: Gets all the info associated with the teams with the match. Summoned players included
      description: Gets all the info associated with the teams with the match. Summoned players included
      tags:
        - Match
        - Team
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: (documentation pending)
  /match/{match_id}/event:
    get:
      summary: Gets the events associated to a specific match
      description: Gets the events associated to a specific match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: List of events that happened during a match
    post:
      summary: registers the events that happened during a match
      description: registers the events that happened during a match
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: events
        in: body
        required: true
        schema:
          $ref: '#/definitions/MatchEvents'
      responses:
        200:
          description: (documentation pending)
  /match/:match_id/team/:team_id/player:
    post:
      summary: registers the starter eleven players for a match for each team
      description: registers the starter eleven players for a match for each team
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: events
        in: body
        required: true
        schema:
          $ref: '#/definitions/initialPlayers'
      responses:
        200:
          description: (documentation pending)
...
  /team/{team_id}/match:
    get:
      summary: Returns all the matches for the team
      description: Returns all the matches for the team
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
      responses:
        200:
          description: An array of past, future matches and the next match for the team
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  CreateMatchRequest:
    title: Match Creation Request
    type: object
    properties:
      number:
        type: integer
      location:
        type: string
      home_team_id:
        type: integer
      visitor_team_id:
        type: integer
      home_team_score:
        type: integer
      visitor_team_score:
        type: integer
      date:
        type: string
      group_id:
        type: integer
      referee_id:
        type: integer
      placeholder_home_team_group:
          type: integer
          description: group_id for the home_team placeholder
      placeholder_home_team_position:
          type: integer
          description: position of the placeholder team according to the group's standing table
      placeholder_visitor_team_group:
          type: integer
          description: group_id for the visitor_team placeholder
      placeholder_visitor_team_position:
          type: integer
          description: position of the placeholder team according to the group's standing table
    required:
    - home_team_id
    - visitor_team_id
    - date
    - location
    - group_id
...
  MatchPlayerTeamRequest:
    title: Match Creation Request
    type: object
    properties:
      position:
        type: string
      number:
        type: integer
      player_id:
        type: integer
    required:
    - player_id
...
  MatchEvents:
    title: MatchEvents
    type: array
    items:
      type: object
      properties:
        event_id:
          type: integer
        player_in:
          type: integer
        player_out:
          type: integer
        instant:
          type: integer
        team_id:
          type: integer
  initialPlayers:
      title: initialPlayers
      type: array
      items:
        type: object
        properties:
          number:
            type: integer
          player_id:
            type: integer
          match_id:
            type: integer
          team_id:
            type: integer
  Error:
    title: error
    type: object
    properties:
      description:
        type: string
...